% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{aa_objective_abs_risk}
\alias{aa_objective_abs_risk}
\title{Asset Allocation objective function - benchmark insensitive}
\usage{
aa_objective_abs_risk(x, cma.ta, cma, risk.budget, OptimizationParameters,
  verbose = FALSE)
}
\arguments{
\item{x}{vector of initial weights at which objective is calculated}

\item{cma.ta}{tax-aware cma}

\item{cma}{cma}

\item{risk.budget}{constraint on pre-tax risk (really a penalty)}

\item{OptimizationParameters}{optimization parameters}

\item{verbose}{FALSE (default) indicates only the objective value is returned; TRUE also returns penalties and other values}

\item{pctassets}{vector with weights of taxed.pct, deferred.pct, exempt.pct.}
}
\value{
list - wtdreturn: weighted return (after-tax, pre-tax), penalties: value of penalties, objectivevalue: objective value
}
\description{
An asset allocation objective function calculating a weighted (w*aftertax + (1-w)*pretax) return less penalties.
Penalties applied to: Non-zero weights below a threshold; risk above a risk.budget threshold; weights outside of box limits;
tracking error above a threshold, the number of classes with non-zero weights below a minimum or above a maximum, and turnover
above a threshold if the curx is supplied.
}
\examples{
aa_objective_abs_risk(x, cma.ta, cma, risk.budget, OptimizationParameters)
}
\seealso{
[create.optimization.parameters()] for creating relevant optimization parameters
}
