% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{aa_objective_min_risk}
\alias{aa_objective_min_risk}
\title{Objective for finding minimum risk portfolio}
\usage{
aa_objective_min_risk(x, cma.ta, cma, OptimizationParameters,
  verbose = FALSE)
}
\arguments{
\item{x}{vector of initial weights at which objective is calculated}

\item{cma.ta}{tax-aware cma}

\item{cma}{cma}

\item{OptimizationParameters}{optimization parameters}

\item{verbose}{FALSE (default) indicates only the objective value is returned; TRUE also returns penalties and other information}
}
\value{
list - wtdreturn: weighted return (after-tax, pre-tax), penalties: value of penalties, objectivevalue: objective value
}
\description{
An asset allocation objective function calculating -100 times pretax risk less penalties.  The -100 factor is to turn a minimization
problem into a maximization one.
Penalties applied to: Non-zero weights below a threshold; risk above a risk.budget threshold; weights outside of box limits;
tracking error above a threshold, the number of classes with non-zero weights below a minimum or above a maximum, and turnover
above a threshold if the curx is supplied.
}
\examples{
aa_objective_min_risk(x, cma.ta, cma, OptimizationParameters)
}
\seealso{
[create.optimization.parameters()] for creating relevant optimization parameters
}
