% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{aa_objective_rel_risk}
\alias{aa_objective_rel_risk}
\title{Asset Allocation objective function - benchmark sensitive}
\usage{
aa_objective_rel_risk(x, cma.ta, cma, wts.bench.ta, wts.bench,
  OptimizationParameters, ClassGroupConstraints, pctassets, curx = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{vector of initial weights at which objective is calculated. Sum of}

\item{cma.ta}{tax-aware cma}

\item{cma}{cma}

\item{wts.bench.ta}{weights of benchmark using cma.ta classes}

\item{wts.bench}{weights of benchmark using cma classes}

\item{OptimizationParameters}{optimization parameters}

\item{ClassGroupConstraints}{class group contraints}

\item{pctassets}{a vector with 3 named elements, taxed.pct, deferred.pct, exempt.pct representing the percentage in each type of account}

\item{curx}{vector of current weights used to calculate turnover}

\item{verbose}{FALSE (default) indicates only the objective value is returned; TRUE also returns penalties and wtd excess return}
}
\value{
list - wtdexcessreturn: weighted excess return (after-tax, pre-tax), penalties: value of penalties, objectivevalue: objective value
}
\description{
An asset allocation objective function calculating a weighted (w*aftertax + (1-w)*pretax) excess return less penalties some of which 
are related to a benchmark (risk less than benchmark, tracking error). Penalties applied to: Non-zero weights below a threshold;
tracking error above a threshold, the number of classes with non-zero weights below a minimum or above a maximum, and turnover
above a threshold if the curx is supplied.
}
\examples{
aa_objective(x, cma.ta, cma, wts.bench.ta, wts.bench, OptimizationParameters, ClassGroupConstraints, pctassets, curx=NULL)
}
\seealso{
[create.optimization.parameters()] for creating relevant optimization parameters
}
