% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{add_port_info}
\alias{add_port_info}
\title{Adds portfolio information to a mix}
\usage{
add_port_info(obj, cma, cma.ta, OptimizationParameters)
}
\arguments{
\item{obj}{A list object containing x (obj$x) which are the aftertax weights}

\item{cma}{capital market assumptions}

\item{cma.ta}{capital market assumptions, tax-adjusted}

\item{OptimizationParameters}{optimization paramters}
}
\value{
The list with additional information including the pretax wts, return and risk stats
}
\description{
Adds portfolio information to a mix
}
\examples{
add_port_info(ojb, cma, cma.ta, OptimizationParameters)
}
