% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATReturn.calc.r
\name{ATReturn.calc}
\alias{ATReturn.calc}
\title{Calculate after-tax expected returns}
\usage{
ATReturn.calc(cma, investor)
}
\arguments{
\item{cma}{Capital market assumption object}

\item{investor}{Investor object}
}
\value{
A list with three numeric vectors: at.return.taxable contains the
  after-tax expected returns for a taxable account, at.return.deferred for a
  deferred account, and at.return.exempt for an exempt account.
}
\description{
An after-tax expected \emph{arithmetic} return is calculated for each asset
class assuming it were held in each of the three account types: taxable,
deferred and exempt. Arithmetic returns are more appropriate inputs into a
mean variance optimizer than geometric.
}
