% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATReturn.calc.r
\name{ATReturn.deferred}
\alias{ATReturn.deferred}
\title{Calculation of after-tax return in a deferred account}
\usage{
ATReturn.deferred(arith.ret, taxROrdInc, taxRState = 0, horizon, risk)
}
\arguments{
\item{arith.ret}{Pre-tax geometric return.  The geometric return is used because
this is the rate at which the asset will grow over time.}

\item{taxROrdInc}{Tax rate on ordinary income}

\item{taxRState}{State income tax rate}

\item{horizon}{Number of year in forecast horizon}

\item{risk}{Standard deviation of asset class.}
}
\value{
The after-tax arithmetic return of an asset invested in taxable account.
}
\description{
The after-tax returns in a deferred account are pretty simple.  An investment
will grow to $(1+pre_tax_geometric_return)^horizon$ and that value will be
subject to tax.  After tax the investor will have this amount less what was
paid in federal and state taxes. We convert this dollar amount into a return
by raising it to the (1/horizon) and subtracting 1. The after-tax geometric
return is then transformed into an arithmetic return.  Most investors think
of the value of their deferred account before taxes.  If someone has $100 in
an IRA, they think it's worth $100 which ignores the deferred tax liability.
By maintaining the fiction that the account today is worth $100, we
understate the after-tax return by dividing an after-tax terminal value with
a before-tax starting value.  Since the final dollars are correct the
allocation is correct which is our goal.
}

