% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATReturn.calc.r
\name{ATReturn.taxable}
\alias{ATReturn.taxable}
\title{Calculation of after-tax return in a taxable account}
\usage{
ATReturn.taxable(yld, growth, valChg, intOrd, intTE, divQual, divOrd,
  turnover, LTCG, STCG, Expense, taxROrdInc, taxRQDiv, taxRLTCG, taxRSTCG,
  taxRState, horizon = 10)
}
\arguments{
\item{yld}{Yield on the investment}

\item{growth}{Growth rate of the income paid on the investment}

\item{valChg}{Change in valuation of the investment}

\item{intOrd}{Percent of income taxed at the ordinary rate}

\item{intTE}{Percent of income not taxed}

\item{divQual}{Percent of income taxed at the qualified dividend rate}

\item{divOrd}{Percent of income taxed at the ordinary dividend rate}

\item{turnover}{Percent of asset sold in the current year. Impacts capital gains.}

\item{LTCG}{Percent of capital gains taxed at long-term rate}

\item{STCG}{Percent of capital gains taxed at short-term rate}

\item{Expense}{Annual expense rate}

\item{taxROrdInc}{Tax rate on ordinary income}

\item{taxRQDiv}{Tax rate on qualified dividends}

\item{taxRLTCG}{Tax rate on long-term capital gains}

\item{taxRSTCG}{Tax rate on short-term capital gains}

\item{taxRState}{State income tax rate}

\item{horizon}{Number of year in forecast horizon}
}
\value{
The after-tax return of an asset invested in taxable account.
}
\description{
The after-tax return calculation in the taxable account is tedious but not
complicated. For each year we calculate the income that would be received.
The income on an investment (as opposed to what may be received by an
investor) grows by a growth rate.  The income is taxed based on its character
(ordinary, qualified, etc) and the investor's tax rate.  Likewise there is
capital gains depending on the growth and the turnover.  We assume the
initial basis equals the initial value (a restriction that could be altered
in future version).  The basis is updated each period according to sales and
purchases.  Looping through all the periods of the horizon we arrive at a
final income. This income along with the valuation change determines the
final value of the asset.  This value is adjusted for a final sale assuming
any gain (loss) is long-term.
}
