% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{calc.ac.wts}
\alias{calc.ac.wts}
\title{Calculate the asset class weights across account types}
\usage{
calc.ac.wts(w, cma.ta)
}
\arguments{
\item{w}{Weights of the assets in a portfolio}

\item{cma.ta}{Tax-aware capital market assumptions.}
}
\value{
value Vector of asset class weights
}
\description{
Tax-aware optimization produces weights for each asset class in each account type.  This function returns the weight of each asset class
across all acount types (e.g. sum of cash in taxable, deferred and exempt accounts)
}

