% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{calc.future.at.value}
\alias{calc.future.at.value}
\title{Calculate the future after-tax value of a portfolio}
\usage{
calc.future.at.value(w, cma.ta, horizon)
}
\arguments{
\item{w}{Weights of the portfolio}

\item{cma.ta}{Tax-aware capital market assumptions}

\item{horizon}{Horizon over which to calculate return}
}
\value{
After-tax value
}
\description{
The returns in the cma are assumed to be arithmetic.  This adjusts them to geometric before compounding.
}
