% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{calc_penalties}
\alias{calc_penalties}
\title{Calculate penalties 
Calculates the penalties used in objective functions
If wts.bench is present, then the penalty will include evaluation of risk relative to the risk of the benchmark
and tracking error.
If risk.budget is a parameter then the penalty will include evaluation of risk over the value of risk budget.
If curx is a parameter then the penalty will include evaluation of turnover.}
\usage{
calc_penalties(...)
}
\arguments{
\item{...}{These parameters must include x - wts; cma.ta, cma, OptimizationParameters, verbose.}
}
\value{
if verbose, a vector of penalties including the total, otherwise just the scalar total.
}
\description{
Calculate penalties 
Calculates the penalties used in objective functions
If wts.bench is present, then the penalty will include evaluation of risk relative to the risk of the benchmark
and tracking error.
If risk.budget is a parameter then the penalty will include evaluation of risk over the value of risk budget.
If curx is a parameter then the penalty will include evaluation of turnover.
}
