% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.add.class}
\alias{cma.add.class}
\title{Add an asset class to a CMA}
\usage{
cma.add.class(cma, class.name, ret, yield, growth, risk, boxmin = 0,
  boxmax = 1, intOrd, intTE, divQual, divOrd, turnover, LTCG, STCG)
}
\arguments{
\item{cma}{Capital market assumption}

\item{class.name}{Name of asset class}

\item{ret}{Expected total return of asset class in decimal (e.g. 0.05 for
five percent)}

\item{yield}{Yield of asset class}

\item{growth}{Growth of asset class}

\item{risk}{Standad deviation of asset class returns}

\item{boxmin}{Minimum weight of asset class. Default 0}

\item{boxmax}{Maximum weight of asset class. Default 1}

\item{intOrd}{The percentage of the yield subject to tax at the ordinary
income tax rate. From 0 to 1.}

\item{intTE}{The percentage of the yield exempt from tax.}

\item{divQual}{The percentage of the yield subject to tax at the qualified
dividend tax rate}

\item{divOrd}{The percentage of the yield subject to tax at the ordinary
dividend tax rate}

\item{turnover}{The percentage of the asset class that is sold each year.}

\item{LTCG}{The percentage of capital gains subject to the long-term capital
gains tax rate.}

\item{STCG}{The percentage of capital gains subject to the short-term capital
gains tax rate.}
}
\value{
A cma with the asset class added.
}
\description{
This function adds an asset class to a cma. Correlation/Covariance data added
separately.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}

