% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.add.constraint}
\alias{cma.add.constraint}
\title{Add constraint to a CMA}
\usage{
cma.add.constraint(constraint, constraint.name, cma)
}
\arguments{
\item{constraint}{a numeric vector with length equal to number of classes.
The dot (scalar) product of this vector and the portfolio weights must be
at least zero to satisfy the constraint.}

\item{constraint.name}{Name of constraint. Default NULL.}

\item{cma}{Capital market assumption object}
}
\value{
A cma with the constraint added.
}
\description{
This function adds a contraint to a cma
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}
