% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.calculate.cov}
\alias{cma.calculate.cov}
\title{Calculates the covariance matrix of a cma from its correlation matrix and
risk vector}
\usage{
cma.calculate.cov(cma)
}
\arguments{
\item{cma}{Capial market assumptions object}
}
\value{
A cma with a covariance matrix computed from its correlation matrix
  and risk vector
}
\description{
Calculates the covariance matrix of a cma from its correlation matrix and
risk vector
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}

