% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.calculate.risk}
\alias{cma.calculate.risk}
\title{Calculates the risk vector of a cma from its covariance matrix}
\usage{
cma.calculate.risk(cma)
}
\arguments{
\item{cma}{Capial market assumptions object}
}
\value{
A cma with the risk vector computed from its covariance matrix.
}
\description{
Calculates the risk vector of a cma from its covariance matrix
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}
