% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.create}
\alias{cma.create}
\title{Create CMA (Capital Market Assumptions)}
\usage{
cma.create(as_of_date = NULL, classes = list(), ret = vector("numeric"),
  yield = vector("numeric"), growth = vector("numeric"),
  risk = vector("numeric"), corr = NULL, cov = NULL, boxMin = 0,
  boxMax = 1, constraints = list(), constraints.names = list())
}
\arguments{
\item{as_of_date}{Date assumptions were created. Default NULL.}

\item{classes}{List of names of asset classes. Default list()}

\item{ret}{Vector of returns, same length as classes. In decimal (e.g. 0.05
for five percent). Default vector("numeric").}

\item{yield}{Vector of yield, same length as classes. In decimal (e.g. 0.05
for five percent). Default vector("numeric").}

\item{growth}{Vector of growth, same length as classes. In decimal (e.g. 0.05
for five percent). Default vector("numeric").}

\item{risk}{Vector of standard deviations, same length as classes. In decimal
(e.g. 0.05 for five percent). Default vector("numeric").}

\item{corr}{Correlation matrix, same length as classes. In decimal (e.g. 0.05
for five percent). Default NULL.}

\item{cov}{Covariance matrix, same length as classes. In decimal (e.g. 0.05
for five percent). Default NULL. If null calculates the covariance matrix
from the correlation matrix and standard deviations if both are provided.}

\item{boxMin}{Vector of minimum wts per asset class. Default is 0 if classes
is not empty otherwise list()}

\item{boxMax}{Vector of maximum wts per asset class. Default is 1 if classes
is not empty otherwise list()}

\item{constraints}{List of constraints.  Each constraint is a numeric vector
with length equal to number of classes. The dot (scalar) product of this
vector and the portfolio weights must be at least zero to satisfy the
constraint. Default List().}

\item{constraints.names}{List of the names of constraints.  Length should
match length of constraints. Default List().}
}
\value{
A list containing items related to the CMA.
}
\description{
This function creates a cma object.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}
