% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.initialize}
\alias{cma.initialize}
\title{Initialize a CMA (Capital Market Assumptions)}
\usage{
cma.initialize(as_of_date = NULL)
}
\arguments{
\item{as_of_date}{Date assumptions were created. Default NULL.}
}
\value{
A cma object with a list of items.
}
\description{
Creates a CMA object.  Expected that add.class and add.constraint will be
called to populate it.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}

