% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.set.corr}
\alias{cma.set.corr}
\title{Add (or replace) the correlation matrix in a cma}
\usage{
cma.set.corr(corr, cma)
}
\arguments{
\item{corr}{Covariance matrix}

\item{cma}{Capial market assumptions object}
}
\value{
A cma with the supplied correlation matrix.
}
\description{
Add (or replace) the correlation matrix in a cma
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}

