% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.set.cov}
\alias{cma.set.cov}
\title{Add (or replace) the covariance matrix in a cma}
\usage{
cma.set.cov(cov, cma)
}
\arguments{
\item{cov}{Covariance matrix}

\item{cma}{Capial market assumptions object}
}
\value{
A cma with the supplied covarance matrix.
}
\description{
Add (or replace) the covariance matrix in a cma
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}
