% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{cma.ta.create}
\alias{cma.ta.create}
\title{Create a tax-aware CMA}
\usage{
cma.ta.create(cma, investor)
}
\arguments{
\item{cma}{A cma object that is not tax-aware}

\item{investor}{An investor object}
}
\value{
A cma.ta object.
}
\description{
Create a tax-aware CMA
}
