% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{combine.class.wts}
\alias{combine.class.wts}
\title{Combine asset class weights from different account types.}
\usage{
combine.class.wts(x, suppress.zero = FALSE)
}
\arguments{
\item{x}{Eff object - efficient frontier}

\item{suppress.zero}{TRUE to suppress output of asset classes with no weights.}
}
\value{
Plot of efficient frontier
}
\description{
Adds the weights, by asset class, of taxable, deferred and exempt accounts
}
