% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{diversification_score}
\alias{diversification_score}
\title{Calculates a measure of diversification}
\usage{
diversification_score(w)
}
\arguments{
\item{w}{Weights of assets in portfolio}
}
\value{
Diversification score.  This score is scaled such that if one equally invested in n assets, the score will be n.
}
\description{
This score is scaled such that if one equally invested in n assets, the score will be n.  Thus higher values indicate greater diversification.
This does not incorporate any correlation among asset classes. It treats each asset class independently.
}

