% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{efficient.frontier}
\alias{efficient.frontier}
\title{Create an efficient frontier}
\usage{
efficient.frontier(cma.ta, n.portfolios = 25)
}
\arguments{
\item{cma.ta}{Tax-aware capital market assumptions}

\item{n.portfolios}{Number of portfolios (points) to produce on the frontier.
Output will have one less point if the minimum return portfolio is
inefficient.}
}
\value{
A matrix. Each row is a point on the frontier.  The first column is
  the return. The second column is the standard deviation. Subsequent columns
  are asset class weights.
}
\description{
This function creates a mean-variant efficient frontier given a set of
capital market assumptions.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}

