% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{find.minmax.return}
\alias{find.minmax.return}
\title{Find Min and Max Return}
\usage{
find.minmax.return(cma.ta)
}
\description{
This function finds the minimum and maximum feasible returns given a set of
 capital market assumptions, including constraints. It is called to find the
 endpoints for creating an efficient frontier. It returns a list with min and
 max items each of which are produced by solveLP. The opt value is the min or
 max respectively. Requires linprog.
}
\details{
@param cma.ta Tax-aware capital market assumptions
 @keywords tax-aware efficient frontier
 @return  value A two item list containing min and max each of which are produced by
   solveLP. The opt value is the min or max return respectively. Arithmetic.
 @export
 @seealso \code{\link[linprog]{solveLP}}
 @keywords find.minmax.return()
}
