% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{find_max_return_portfolio}
\alias{find_max_return_portfolio}
\title{Find Max Return portfolios for a risk budget}
\usage{
find_max_return_portfolio(...)
}
\arguments{
\item{...}{Parameters to pass to aa_objective_abs_risk except for x}
}
\value{
list of values
}
\description{
Find Max Return portfolios for a risk budget
}
\examples{
find_max_return_portfolio(cma.ta=cma.ta, cma=cma, risk.budget=rb, OptimizationParameters=OptimizationParameters, verbose=FALSE)
}
