% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{find_min_risk_portfolio}
\alias{find_min_risk_portfolio}
\title{Find the minimum risk portfolio}
\usage{
find_min_risk_portfolio(...)
}
\arguments{
\item{...}{parameters to pass to the aa_objective_min_risk function}
}
\value{
Results of maximize_pairwise function
}
\description{
Find the minimum risk portfolio
}
\examples{
find_min_risk_portfolio(0.03, cma.ta, cma, 0.03, OptimizationParameters, FALSE)
}
\seealso{
[maximize_pairwise()] for more detail on return values.
}
