% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{Get10YrBEInflationRate}
\alias{Get10YrBEInflationRate}
\title{Retrieves the 10 year breakeven inflation rate from the St. Louis Fed (FRED)}
\usage{
Get10YrBEInflationRate(dt = format(Sys.Date(), "\%Y-\%m-01"))
}
\arguments{
\item{dt}{Character representing date for which the rate is to be retrieved. YYYY-MM-DD. Default is system dae.}
}
\value{
A list with a cma with date and rate items. The date is the date the statistic was calculated.
}
\description{
See \url{https://research.stlouisfed.org/fred2/series/T10YIE} for more info
Uses the Quandl package to retrieve the data.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier.}
