% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investor.create.r
\name{investor.create}
\alias{investor.create}
\title{Create Investor}
\usage{
investor.create(value.taxable, value.deferred, value.exempt, horizon = 10,
  taxrate.state = 0, taxrate.ordinc = 0, taxrate.LTCG = 0,
  taxrate.STCG = 0, taxrate.qualdiv = 0, taxrate.surcharge = 0.038,
  income = NULL)
}
\arguments{
\item{value.taxable}{Value in taxable account}

\item{value.deferred}{Value in deferred account}

\item{value.exempt}{Value in exempt account.}

\item{horizon}{Time horizon in years. Default is 10.}

\item{taxrate.state}{State income tax rate. Default is 0.}

\item{taxrate.ordinc}{Federal tax rate on ordinary income. Default is 0.}

\item{taxrate.LTCG}{Federal tax rate on long-term capital gains. Default is
0.}

\item{taxrate.STCG}{Federal tax rate on short-term capital gains. Default is
0.}

\item{taxrate.qualdiv}{Federal tax rate qualified dividends. Default is 0.}

\item{taxrate.surcharge}{Net investment income tax rate (section 1411 of
IRC). Default is 0.038.}

\item{income}{Income level. Might be used for looking up tax rates. Default
is NULL}
}
\value{
A list containing items related to the investor. This package
  internally calculate income and capital gain rates that combine federal,
  state and surcharge taxes.
}
\description{
This function creates an investor object.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}

