% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investor.create.r
\name{investor.pctassets}
\alias{investor.pctassets}
\title{Create a vector of the percent of pretax assets}
\usage{
investor.pctassets(x)
}
\arguments{
\item{x}{investor object}
}
\value{
vector of numeric
}
\description{
Create a vector of the percent of pretax assets
}
\examples{
invstor.pctassets(investor)
}
