% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investor.create.r
\name{investor.value}
\alias{investor.value}
\title{Initial value in investor accounts}
\usage{
investor.value(investor, tax = "after")
}
\arguments{
\item{investor}{Investor object}

\item{tax}{"after" for after-tax values, otherwise before-tax}
}
\value{
Sum of taxable, deferred and exempt values
}
\description{
Initial value in investor accounts
}
