% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbypairwiseiter.r
\name{make_alloc_tables}
\alias{make_alloc_tables}
\title{Make allocation tables
Give a set of weights, the function returns two tables.  The weights in the allocationacrossaccount table
will sum to 100.  Each value represents the allocation of the entire portfolio.  The weights in the
allocationbyaccount shows the weight within each account.  The weights in each account will sum to 100.
These are useful for display.}
\usage{
make_alloc_tables(wts, cma.ta)
}
\arguments{
\item{wts}{Weights across assets}

\item{cma.ta}{capital market assumptions, tax-advantaged}
}
\value{
list with two tables allocationacrossaccount and allocationbyaccount
}
\description{
Make allocation tables
Give a set of weights, the function returns two tables.  The weights in the allocationacrossaccount table
will sum to 100.  Each value represents the allocation of the entire portfolio.  The weights in the
allocationbyaccount shows the weight within each account.  The weights in each account will sum to 100.
These are useful for display.
}
\examples{
make_alloc_tables(wts, cma.ta)
}
