% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{optimize.target.return}
\alias{optimize.target.return}
\title{Find the portfolio with the minimum standard deviation given a target return
and a set of capital market assumptions}
\usage{
optimize.target.return(target.ret, cma.ta, tol = 4e-05)
}
\arguments{
\item{target.ret}{The target return.  Should be a feasible return.}

\item{cma.ta}{Tax-aware capital market assumptions}

\item{tol}{Tolerance for constraints.}
}
\value{
The optimum produced by solve.QP from quadprog. The solution value is the set of weights.
}
\description{
This function is used to create an efficient frontier which iterates across a
set of target returns. Requires quadprog
}
\seealso{
\code{\link[quadprog]{solve.QP}}
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
\keyword{return}
\keyword{target}

