% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{plot.eff}
\alias{plot.eff}
\title{Plot efficient frontier}
\usage{
\method{plot}{eff}(x, ...)
}
\arguments{
\item{x}{Efficient frontier object}

\item{...}{}
}
\value{
plot
}
\description{
Plot efficient frontier
}
