% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{portrisk}
\alias{portrisk}
\title{Calculate the standard deviation of a portfolio}
\usage{
portrisk(w, cma)
}
\arguments{
\item{w}{Weights of the assets in a portfolio}

\item{cma}{Capital market assumptions.  Must have a cov item that is a square
matrix each dimension matching the length of w}
}
\value{
value representing the standard deviation of the portfolio
}
\description{
This function calls the portrisk function
}
