% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.initialize.r
\name{print.cma.ta}
\alias{print.cma.ta}
\title{Print cma.ta object}
\usage{
\method{print}{cma.ta}(cma.ta, kable = TRUE, ...)
}
\arguments{
\item{cma.ta}{cma.ta object}

\item{kable}{TRUE returns result of the knitr kable function.}

\item{...}{Additional print parameters}
}
\value{
object to print
}
\description{
Print cma.ta object
}
