% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{print.eff}
\alias{print.eff}
\title{Print Efficient Frontier}
\usage{
\method{print}{eff}(x, content = c("", "brief", "combined", "detailed"),
  kable = TRUE, transpose = FALSE, ...)
}
\arguments{
\item{x}{Efficient frontier object}

\item{content}{Level of detail of output.  'brief' only shows risk and
return, 'combined' adds base asset classes, 'detailed' includes asset
classes at account level.}

\item{kable}{TRUE returns result of the knitr kable function.}

\item{transpose}{TRUE returns a transpose of the output}

\item{...}{Additional named values.}
}
\value{
print object
}
\description{
Print Efficient Frontier
}
