% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investor.create.r
\name{print.investor}
\alias{print.investor}
\title{Print investor}
\usage{
\method{print}{investor}(x, ...)
}
\arguments{
\item{x}{Investor}

\item{...}{Additional print parameters}
}
\value{
Prints
}
\description{
Print investor
}
