% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.cma}
\alias{rafi.cma}
\title{Load RAFI data into a cma}
\usage{
rafi.cma(rafi.data.loc, acnametable = "acname_table.csv",
  inflation.rate = Get10YrBEInflationRate()$rate)
}
\arguments{
\item{rafi.data.loc}{Folder in which the CSV files are located}

\item{acnametable}{Name of csv file containing asset class details}

\item{inflation}{rate Default calls the function
\code{\link{Get10YrBEInflationRate}}}
}
\value{
A list containing the as_of_date of the assumptions, the nclasses
  (number of classes), ac.data which is a table in which each row represents
  and asset class and columns contain return, risk and tax information.  This
  table also includes box and custom constraints There is also corr the
  correlation matrix, and/or cov for the covariance matrix and nconstraints for the number of
  constraints.
}
\description{
This function reads a set of returns and correlations from CSV files
downloaded from the RAFI website.  Since RAFI may change its file formats,
this function may not work as expected in the future. This data takes the raw
RAFI estimates and massages the.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}

