% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.cma.v2}
\alias{rafi.cma.v2}
\title{Load RAFI data into a cma
Version 2 uses RAFI file format which was adopted around May 2017
This new format includes a nominal return from which the inflation rate 
may be inferred.}
\usage{
rafi.cma.v2(rafi.data.loc, acnametable = "acname_table.xlsx",
  xls.file.name = "Asset-Allocation-Interactive-Data.xlsx")
}
\arguments{
\item{rafi.data.loc}{Folder in which the CSV files are located}

\item{acnametable}{Name of XLSX file containing asset class details}

\item{xls.file.name}{name of XLSX file with RAFI assumptions}
}
\value{
A cma object which is a list containing the as_of_date of the
  assumptions, the nclasses (number of classes), ac.data which is a table in
  which each row represents and asset class and columns contain return, risk
  and tax information.  This table also includes box and custom constraints
  There is also corr the correlation matrix, and/or cov for the covariance
  matrix and nconstraints for the number of constraints.
}
\description{
This function reads a set of returns and correlations from XLSX files 
downloaded from the RAFI website.  Since RAFI may change its file formats, 
this function may not work as expected in the future.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
