% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.data.check}
\alias{rafi.data.check}
\title{Check RAFI data}
\usage{
rafi.data.check(rafi.data.loc, acnametable)
}
\arguments{
\item{rafi.data.loc}{folder in which the CSV files are located}

\item{acnametable}{Name of csv file containing asset class details}
}
\value{
A list with errors or and empty list is no errors
}
\description{
This function checks the 2 CSV files from RAFI.
It checks that the names of the assets classes are defined in the acnametable
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
