% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.load}
\alias{rafi.load}
\title{Load RAFI data}
\usage{
rafi.load(rafi.data.loc, acnametable = "acname_table.csv")
}
\arguments{
\item{rafi.data.loc}{folder in which the CSV files are located}

\item{acnametable}{Name of csv file containing asset class details}
}
\value{
A list with four items: ret is the return data, corr is the correlation matrix and
as_of_date is the date of the assumptionsfrom the csv file, nclasses is the number of classes
  correlation data
}
\description{
This function reads a set of returns and correlations from CSV files
downloaded from the RAFI website.  Since RAFI may change its file formats,
this function may not work as expected in the future.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
