% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.load.v1}
\alias{rafi.load.v1}
\title{Load RAFI data}
\usage{
rafi.load.v1(rafi.data.loc, acnametable = "acname_table.csv",
  file.ret = "core_asset_class_expected_returns.csv",
  file.corr = "core_asset_class_correlations_forecasted.csv")
}
\arguments{
\item{rafi.data.loc}{folder in which the CSV files are located}

\item{acnametable}{Name of csv file containing asset class details}

\item{file.ret}{CSV file name with return assumptions}

\item{file.corr}{CSV file name with correlation assumptions}
}
\value{
A list with four items: ret is the return data, corr is the correlation matrix and
as_of_date is the date of the assumptionsfrom the csv file, nclasses is the number of classes
  correlation data
}
\description{
This function reads a set of returns and correlations from CSV files
downloaded from the RAFI website.  Since RAFI may change its file formats,
this function may not work as expected in the future.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
