% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{RAFI.read.csv}
\alias{RAFI.read.csv}
\title{Read RAFI csv files}
\usage{
RAFI.read.csv(rafi.data.loc)
}
\arguments{
\item{rafi.data.loc}{folder in which the CSV files are located}
}
\value{
A list with two items: ret is the return data, corr is the
  correlation data
}
\description{
This function reads a set of returns and correlations from CSV files
downloaded from the RAFI website.  No processing is done. This function
allows one to see the data in an unprocessed form.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}

