% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.read.csv.v1}
\alias{rafi.read.csv.v1}
\title{Read RAFI csv files}
\usage{
rafi.read.csv.v1(rafi.data.loc,
  file.ret = "core_asset_class_expected_returns.csv",
  file.corr = "core_asset_class_correlations_forecasted.csv")
}
\arguments{
\item{rafi.data.loc}{folder in which the CSV files are located}

\item{file.ret}{CSV file name with return assumptions}

\item{file.corr}{CSV file name with correlation assumptions}
}
\value{
value A list with two items: ret is the return data, corr is the 
  correlation data
}
\description{
This function reads a set of returns and correlations from CSV files 
downloaded from the RAFI website.  No processing is done. This function 
allows one to see the data in an unprocessed form.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
