% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.read.xls.v2}
\alias{rafi.read.xls.v2}
\title{Read RAFI XLS file
 
This function reads a set of returns and correlations from XLS file 
downloaded from the RAFI website.  No processing is done. This function 
allows one to see the data in an unprocessed form.}
\usage{
rafi.read.xls.v2(rafi.data.loc,
  xls.file.name = "Asset-Allocation-Interactive-Data.xlsx")
}
\arguments{
\item{rafi.data.loc}{folder in which the files are located}

\item{xls.file.name}{name of XLSX file}
}
\value{
value A list with two items: ret is the return data, corr is the 
  correlation data
}
\description{
Read RAFI XLS file
 
This function reads a set of returns and correlations from XLS file 
downloaded from the RAFI website.  No processing is done. This function 
allows one to see the data in an unprocessed form.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
