% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.frontier.r
\name{resample.target.risk}
\alias{resample.target.risk}
\title{Use resampling to find an efficient portfolio}
\usage{
resample.target.risk(target.risk, cma.ta, n.samples = 100, thresh = 0,
  tol = 0.00004)
}
\arguments{
\item{target.risk}{The target risk for which to find an efficient portfolio.}

\item{cma.ta}{Tax-aware capital market assumptions}

\item{n.samples}{The number of samples for the resampling.}

\item{thresh}{The minimum weight for an asset class (if the assumptions do
not specify another minimum)}

\item{tol}{Tolerance to pass to Optimize Target Return}
}
\value{
A list with two items.  The first item w is the weights of the
  efficient portfolio. The second item is mat which is the matrix of
  resampled solutions.  Each row is a sample. The first two columns are
  return and risk respectively.  The remainding columns are the weights.  The
  w item is the column means of the weights.
}
\description{
This function finds an efficient frontier given a target standard deviation
and a set of capital market assumptions. Resampling reduces the impact of
estimation error in the mean-variance process.  The process starts with a set
of assumptions. These are used to generate a set of random returns (a
sample). The return and covariance of this sample is computed and an
efficient point is created. This sampling is repeated.  The average
allocations of the samples is the result. This function allows the user to
set a minimum threshold for an asset class.  If an asset class's weight fails
to meet the threshold, the class's maximum weight is set to zero.  The
resampling is repeated.  The threshold is ignored for asset classes with
minimum weights less than the threshold (for example, if one sets a 1%
minimum weight to cash and a threshold of 2.5%, then a 1% cash weighting will
be allowed.)
}
