% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{ret_SqRootofT}
\alias{ret_SqRootofT}
\title{Return using the square-root-of-T rule}
\usage{
ret_SqRootofT(p, ExpRetAnnual, SD, T, method = "RAFI")
}
\arguments{
\item{p}{probability of doing worse than the result}

\item{ExpRetAnnual}{Annualized expected return}

\item{SD}{Annualized standard deviation}

\item{T}{Time in years}

\item{method}{(unused)}
}
\value{
value Annualized return with at least a probability of p of occurring.
}
\description{
This calculation follows that described on page 16 of RAFI AA-Asset_Class-Risk.pdf document.
}
\examples{
ret_SqRootofT(0.025,1.3,14.3,10)
}
