% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{portret}
\alias{portret}
\title{Calculate the return of a portfolio}
\usage{
portret(weights, cma)
}
\arguments{
\item{cma}{Capital market assumptions.  Must have a ret item with a length
the length of w}

\item{w}{Weights of the assets in a portfolio}
}
\value{
value representing the variance of the portfolio
}
\description{
This is the expected return given the capital market assumptions
}
