% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portstats.r
\name{portvar}
\alias{portvar}
\title{Calculate the variance of a portfolio}
\usage{
portvar(w, cma)
}
\arguments{
\item{w}{Weights of the assets in a portfolio}

\item{cma}{Capital market assumptions.  Must have a cov item that is a square
matrix each dimension matching the length of w}
}
\value{
value representing the variance of the portfolio
}
\description{
This function is called by the portrisk function
}

