% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafi.r
\name{rafi.cma}
\alias{rafi.cma}
\title{Load RAFI data into a cma}
\usage{
rafi.cma(version = "v2", rafi.data.loc,
  acnametable = "acname_table.xlsx",
  xls.file.name = "Asset-Allocation-Interactive-Data.xlsx",
  file.ret = "core_asset_class_expected_returns.csv",
  file.corr = "core_asset_class_correlations_forecasted.csv",
  inflation.rate = NULL)
}
\arguments{
\item{version}{Version (either v1 or v2). Use v1 if RAFI files are in CSV format
through April 2017.  From May 2017 they use a XLSX format.}

\item{rafi.data.loc}{Folder in which the RAFI input files are located}

\item{acnametable}{Name of csv or xlsx file containing asset class details}

\item{xls.file.name}{Name of XLSX RAFI files}

\item{file.ret}{CSV file name with return assumptions}

\item{file.corr}{CSV file name with correlation assumptions}

\item{inflation.rate}{Default calls the function
\code{\link{Get10YrBEInflationRate}}}
}
\value{
A cma object which is a list containing the as_of_date of the
  assumptions, the nclasses (number of classes), ac.data which is a table in
  which each row represents and asset class and columns contain return, risk
  and tax information.  This table also includes box and custom constraints
  There is also corr the correlation matrix, and/or cov for the covariance
  matrix and nconstraints for the number of constraints.
}
\description{
This function reads a set of returns and correlations
downloaded from the RAFI website. Combined with assumptions about
the investor, a cma object is returned representing capital
market assumptions.  RAFI files are in CSV format
 through April 2017.  From May 2017 they use a XLSX format.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
